Moved and improved CMake auto pkg build/install.
Changing course from here.  Spending too much time writing the build
system and not enough time using it.  Will change structure of
directories to support multiple installed versions at once, and start
doing a lot less 'autogit' behavior.  Likely moving back toward the
original Python scripts used to find modules and download them.  CMake
hasn't worked out as hoped for that purpose.  add_subdirectory() works
great, but imposes a structre we don't want.  ExternalProject_Add() is
nice in theory, but tries to do too many things automatically.  Can't
control it in the way we'd want, and it's setup to try updating git
ExternalProject to HEAD every time a user of that project builds.
Sounds slow and is totally unwanted if a particular version is desired.
At which points commands in CMake for doing things like
updating/patching can be set to the empty string, but then so much of
ExternalProject's behavior is gutted that it's still not worth it.